requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-core
azure-storage-blob
alabaster==0.7.12
appdirs==1.4.4
argh==0.26.2
argon2-cffi==20.1.0
astroid==2.4.2
astropy==4.2
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.4
Babel==2.9.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
black==19.10b0
bleach==3.2.1
bokeh==2.2.3
Bottleneck==1.3.2
brotlipy==0.7.0
bs4==0.0.1
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
cryptography==3.3.1
cycler==0.10.0
cytoolz==0.11.0
dask==2020.12.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
dill==0.3.3
distributed==2020.12.0
docutils==0.16
entrypoints==0.3
fast-histogram==0.9
flake8==3.8.4
fsspec==0.8.3
future==0.18.2
glue-core
glue-vispy-viewers
glueviz
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-trans-new==1.1.9
googleapis-common-protos==1.52.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
h5py==2.10.0
HeapDict==1.0.1
hpack==3.0.0
hstspreload==2020.11.21
httpcore==0.9.1
httplib2==0.18.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imagecodecs==2020.5.30
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==2.0.0
intervaltree==3.1.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
keyring==21.4.0
kiwisolver==1.3.0
lazy-object-proxy==1.4.3
locket==0.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
mpl-scatter-density==0.7
msgpack==1.0.1
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.6
numpy==1.19.2
numpydoc==1.1.0
oauthlib==3.1.0
olefile==0.46
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.7.0
partd==1.1.0
pathspec==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pip==20.3.3
plotly==4.14.1
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
psutil==5.7.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
pyerfa==1.7.1.1
pyflakes==2.2.0
Pygments==2.7.3
pylint==2.6.0
pyls-black==0.4.6
pyls-spyder==0.3.0
PyNaCl==1.4.0
PyOpenGL==3.1.1a1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2020.4
PyWavelets==1.1.1
# pywin32
# pywin32-ctypes==0.2.0
# pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==20.0.0
QDarkStyle==2.8.1
QtAwesome==1.0.1
qtconsole==4.7.7
QtPy==1.9.0
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
retrying==1.3.3
rfc3986==1.4.0
rope==0.18.0
rsa==4.6
# Rtree==0.9.4
schedule==0.6.0
scikit-image==0.17.2
scipy==1.5.2
selenium==3.141.0
Send2Trash==1.5.0
setuptools==51.1.0
six==1.15.0
sniffio==1.2.0
snowballstemmer==2.0.0
sortedcontainers==2.3.0
soupsieve==2.1
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.2.0
spyder-kernels==1.10.0
tblib==1.7.0
terminado==0.9.2
testpath==0.4.4
three-merge==0.1.1
tifffile==2020.12.8
toml==0.10.1
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
ujson==4.0.1
uritemplate==3.0.1
urllib3==1.25.11
watchdog==0.10.4
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.36.2
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
xlrd==2.0.1
yapf==0.30.0
zict==2.0.0
zipp==3.4.0

